refactor(InputView): update validate method, remove Validator class
